moved addMeta, addVars, roundValues, reformatTime, reformatLon to write.py
